backend/intelligent_chat/requirements.txt
# FastAPI and core dependencies
fastapi==0.116.1
uvicorn==0.35.0
uvloop==0.21.0
httptools==0.6.4
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic==2.11.7
pydantic_core==2.33.2
pydantic-settings==2.7.1

# Database
asyncpg==0.30.0
SQLAlchemy==2.0.41
psycopg==3.2.9
psycopg-pool==3.2.6

# AI/ML Libraries
boto3==1.39.8
botocore==1.39.8
pinecone-client==4.1.2
langchain==0.3.27
langchain-anthropic==0.3.18
langchain-aws==0.2.30
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.0
langgraph-checkpoint==2.1.1
langgraph-checkpoint-postgres==2.0.23
langgraph-prebuilt==0.6.0
langgraph-sdk==0.2.0
langsmith==0.4.8
langmem==0.0.29
trustcall==0.0.39
openai==1.97.1

# Document processing
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==11.1.0
aiofiles==24.1.0

# Text processing
nltk==3.9.1
tiktoken==0.9.0
textblob==0.17.1

# Redis
redis==6.2.0
hiredis==3.2.1

# Utilities
requests==2.32.4
aiohttp==3.12.15
tenacity==9.1.2
PyJWT==2.10.1
bcrypt==4.3.0

# Data processing
pandas==2.3.1
numpy==2.3.2

